+++ String Algorithms/Pattern Positions.cpp
@@ -82,7 +82,7 @@ int query_idx(){
 
 int main(){
     scanf(" %s %d", S, &K);
-    N = strlen(S);
+    N = (int) strlen(S);
 
     init();
     for(int i = 0; i < N; i++)
@@ -91,7 +91,7 @@ int main(){
 
     for(int i =  0; i < K; i++){
         scanf(" %s", T);
-        M = strlen(T);
+        M = (int) strlen(T);
         printf("%d\n", query_idx());
     }
 }
